Switch from js-yaml to yamljs
Turns out, per #3, that js-yaml isn't browserify-friendly. It appears that
yamljs is. However, this commit currently causes a test failure:

```
  6 passing (33ms)
  1 failing

  1) validate should return an array of errors if the YAML fails to parse:
     AssertionError: expected undefined not to be empty
      at Context.<anonymous> (test/validator-test.js:48:32)
```

The following payload should fail because strings that contain `:` should be
wrapped in quotes:

```js
      var yamlContents = 'lyric: watch out boy: she\'ll chew you up!',
```

But validation passes because yamljs is a little too smart, and will figure
out how to wrap such malformed strings:

```js
{ lyric: 'watch out boy: she\'ll chew you up!' }
```

Question: Accept this robust behavior, and find another parse error to use in
the test, or try finding another library that is as strict as js-yaml?